% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{all_indices}
\alias{all_indices}
\title{All indices combined}
\usage{
all_indices(Yp, Ys, Mp, Ms)
}
\arguments{
\item{Yp}{Yield under control condition}

\item{Ys}{Yield under stress condition}

\item{Mp}{Mean yield of all the genotypes under control condition}

\item{Ms}{Mean yiels of all the genotyps under Stress condition}
}
\value{
Indices Combined
}
\description{
Function to all the indices related to biomass/ yield under different growth conditions
}
\examples{
Mp <- mean(yield_data$Yp)
Ms <- mean(yield_data$Ys)
Yp <- yield_data$Yp
Ys <- yield_data$Ys
all_indices(Yp, Ys, Mp, Ms)
}
