% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{ATI}
\alias{ATI}
\title{Abiotic Tolerance Index
Calculate abiotic tolerance index according to Moosavi et al. (2008)}
\usage{
ATI(Yp, Ys, Mp, Ms)
}
\arguments{
\item{Yp}{Yield under control condition}

\item{Ys}{Yield under Stress condition}

\item{Mp}{Mean yield of all the genotypes under stress condition}

\item{Ms}{Mean yield of all the genotypes under control condition}
}
\value{
ATI
}
\description{
Abiotic Tolerance Index
Calculate abiotic tolerance index according to Moosavi et al. (2008)
}
\examples{
ATI(500, 350, 400, 300)
}
\references{
Moosavi SS, Samadi YB, Naghavi MR, Zali AA, Dashti H, Pourshahbazi A (2008) Introduction of new indices to identify relative drought tolerance and resistance in wheat genotypes. Desert. 12: 165-178.
}
