% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{calculate_PEG_6000}
\alias{calculate_PEG_6000}
\title{Calculate Polyethylene glycol (PEG) 6000 requirement}
\usage{
calculate_PEG_6000(C, bar)
}
\arguments{
\item{C}{Temperature of solution in degree centigrade}

\item{bar}{Water potential in bars}
}
\value{
PEG6000 required
}
\description{
Calculate Amount of PEG6000 required to reach desired water potential at given temperature
}
\examples{
calculate_PEG_6000(25, -4)
}
\references{
Michel, B. E., & Kaufmann, M. R. (1973). The osmotic potential of polyethylene glycol 6000. Plant physiology, 51(5), 914-916.
}
