% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{caro_total}
\alias{caro_total}
\title{Total carotenoids content}
\usage{
caro_total(A470, A663, A646, fresh_weight)
}
\arguments{
\item{A470}{Absorbance at 470nm}

\item{A663}{Absorbance at 663nm}

\item{A646}{Absorbance at 646nm}

\item{fresh_weight}{Fresh weight of the sample used in grams}
}
\value{
Carotenoids concentration in µg/ml
}
\description{
Calculate total carotenoids using Method by Lichtenthaler (1987)
}
\examples{
caro_total(0.7, 0.041, 0.025, 1)
}
\references{
Lichtenthaler, H. K. (1987). Chlorophylls and carotenoids: pigments of photosynthetic biomembranes. In Methods in enzymology (Vol. 148, pp. 350-382). Academic Press.
}
