% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{chl_a}
\alias{chl_a}
\title{Chlorophyll'a' Concentration by Arnon method
Calculates Chlorophyll a Concentration according to Arnon(1949) method}
\usage{
chl_a(A663, A645, v, w)
}
\arguments{
\item{A663}{Absorbance at 663nm}

\item{A645}{Absorbance at 645nm}

\item{v}{Final volume of solvent used in ml}

\item{w}{Fresh weight of the sample used in grams}
}
\value{
Chlorophyll a in mg/g of fresh weight
}
\description{
Chlorophyll'a' Concentration by Arnon method
Calculates Chlorophyll a Concentration according to Arnon(1949) method
}
\examples{
chl_a(0.025, 0.041, 15, 1)
}
\references{
Arnon, D. I. (1949). Copper enzymes in isolated chloroplasts. Polyphenoloxidase in Beta vulgaris. Plant physiology, 24(1), 1. \doi{10.1104/pp.24.1.1}
}
