% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{chl_b}
\alias{chl_b}
\title{Chlorophyll b concentration
Calculates Chlorophyll b Concentration according to Arnon(1949) method}
\usage{
chl_b(A645, A663, v, w)
}
\arguments{
\item{A645}{Absorbance at 645nm}

\item{A663}{Absorbance at 663nm}

\item{v}{Final volume of solvent used in ml}

\item{w}{Fresh weight of the sample used in grams}
}
\value{
Chlorophyll b in mg/g of fresh weight
}
\description{
Chlorophyll b concentration
Calculates Chlorophyll b Concentration according to Arnon(1949) method
}
\examples{
chl_b(0.041, 0.025, 15, 1)
}
\references{
Arnon, D. I. (1949). Copper enzymes in isolated chloroplasts. Polyphenoloxidase in Beta vulgaris. Plant physiology, 24(1), 1. \doi{10.1104/pp.24.1.1}
}
