% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{chl_total}
\alias{chl_total}
\title{Total chlorophyll (a+b) concentration
Calculate Total chlorophyll (a+b) concentration using method by Arnon (1949)}
\usage{
chl_total(A645, A663, v, w)
}
\arguments{
\item{A645}{Absorbance at 645nm}

\item{A663}{Absorbance at 663nm}

\item{v}{Final volume of solvent used in ml}

\item{w}{Fresh weight of the sample used in grams}
}
\value{
Total chlorophyll (a+b) in mg/g of fresh weight
}
\description{
Total chlorophyll (a+b) concentration
Calculate Total chlorophyll (a+b) concentration using method by Arnon (1949)
}
\examples{
chl_total(0.041, 0.025, 15, 1)
}
\references{
Arnon, D. I. (1949). Copper enzymes in isolated chloroplasts. Polyphenoloxidase in Beta vulgaris. Plant physiology, 24(1), 1. \doi{10.1104/pp.24.1.1}
}
