% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{DRI}
\alias{DRI}
\title{Drought resistant index (DRI)
The genotype with high values of this index will be more suitable for drought stress condition}
\usage{
DRI(Yp, Ys)
}
\arguments{
\item{Yp}{Yield under control condition}

\item{Ys}{Yield under stress condition}
}
\value{
DRI
}
\description{
Drought resistant index (DRI)
The genotype with high values of this index will be more suitable for drought stress condition
}
\examples{
DRI(500, 350)
}
