% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{gmp}
\alias{gmp}
\title{Geometric mean productivity (GMP) by Fernandez (1992)}
\usage{
gmp(Yp, Ys)
}
\arguments{
\item{Yp}{Yield under control condition}

\item{Ys}{Yield under stress condition}
}
\value{
GMP
}
\description{
The genotype with high values of this index will be more desirable
}
\examples{
gmp(5, 3)
}
\references{
Fernandez, G. C. (1993). Effective selection criteria for assessing plant stress tolerance.
}
