% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{Golden_mean}
\alias{Golden_mean}
\title{Golden Mean
Calculates Golden mean value using Moradi et al.,(2012)}
\usage{
Golden_mean(Yp, Ys)
}
\arguments{
\item{Yp}{Yield under control condition}

\item{Ys}{Yield under stress condition}
}
\value{
GM
}
\description{
Golden Mean
Calculates Golden mean value using Moradi et al.,(2012)
}
\examples{
 Golden_mean(500,350)
}
\references{
Moradi H, Akbari GA, Khorasani SK, Ramshini HA (2012) Evaluation of drought tolerance in corn (Zea Mays L.) new hybrids with using stress tolerance indices. Eur J Sustain Dev 1. (3): 543-560
}
