% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{HAM}
\alias{HAM}
\title{Harmonic Mean}
\usage{
HAM(Yp, Ys)
}
\arguments{
\item{Yp}{Yield under control condition}

\item{Ys}{Yield under stress condition}
}
\value{
Harmonic mean
}
\description{
Harmonic Mean
}
\examples{
HAM(500, 350)
}
\references{
Schneider, K. A., Rosales‐Serna, R., Ibarra‐Perez, F., Cazares‐Enriquez, B., Acosta‐Gallegos, J. A., Ramirez‐Vallejo, P., ... & Kelly, J. D. (1997). Improving common bean performance under drought stress. Crop science, 37(1), 43-50.
}
