% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{mp_index}
\alias{mp_index}
\title{Mean productivity Index- by Hossain etal., (1990)}
\usage{
mp_index(Yp, Ys)
}
\arguments{
\item{Yp}{Yield under control condition}

\item{Ys}{Yield under stress condition}
}
\value{
Mean productivity Index
}
\description{
The genotype with high values of this index will be more desirable
}
\references{
Hossain, A. B. S., Sears, R. G., Cox, T. S., & Paulsen, G. M. (1990). Desiccation tolerance and its relationship to assimilate partitioning in winter wheat. Crop Science, 30(3), 622-627.
}
