% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{peg_6000}
\alias{peg_6000}
\title{Water potential of Polyethylene glycol (PEG) 6000}
\usage{
peg_6000(peg, C)
}
\arguments{
\item{peg}{Amount PEG600O in grams}

\item{C}{Temperature of the solution in degree centigrade}
}
\value{
Water potential in bars
}
\description{
Calculate the corresponding water potential of PEG6000 when dissolved in 1l of water
}
\examples{
peg_6000(20, 25)
}
\references{
Michel, B. E., & Kaufmann, M. R. (1973). The osmotic potential of polyethylene glycol 6000. Plant physiology, 51(5), 914-916.
}
