% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{R_drought_index}
\alias{R_drought_index}
\title{Relative Drought Index
Calculates relative drought index according to Fisher and Wood (1979)}
\usage{
R_drought_index(Yp, Ys, Mp, Ms)
}
\arguments{
\item{Yp}{Yield under control condition}

\item{Ys}{Yield under stress condition}

\item{Mp}{Mean Yield of all the genotypes under control Condition}

\item{Ms}{Mean Yield of all the genotypes under stress Condition}
}
\value{
RDI
}
\description{
Relative Drought Index
Calculates relative drought index according to Fisher and Wood (1979)
}
\examples{
R_drought_index(500, 350, 400, 300)
}
\references{
Fischer RA, Wood JT (1979) Drought resistance in spring wheat cultivars III. Yield association with morphological traits. Aust J Agr Res. 30: 1001-1020
}
