% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{ss_index}
\alias{ss_index}
\title{Stress susceptibility index (SSI) by #Fischer and Maurer (1978)
The genotype with high SSI < 1 are more resistant to drought stress conditions}
\usage{
ss_index(Yp, Ys, Ms, Mp)
}
\arguments{
\item{Yp}{Yield under control condition}

\item{Ys}{Yield under Stress condition}

\item{Ms}{Mean yield of all the genotypes under control condition}

\item{Mp}{Mean yield of all the genotypes under stress condition}
}
\value{
SSI
}
\description{
Stress susceptibility index (SSI) by #Fischer and Maurer (1978)
The genotype with high SSI < 1 are more resistant to drought stress conditions
}
\examples{
ss_index(500, 350, 450, 370)
}
\references{
Fischer, R. A., & Maurer, R. (1978). Drought resistance in spring wheat cultivars. I. Grain yield responses. Australian Journal of Agricultural Research, 29(5), 897-912.
}
