% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{tol_index}
\alias{tol_index}
\title{Tolerance index -TOL by Hossain etal., (1990)}
\usage{
tol_index(Yp, Ys)
}
\arguments{
\item{Yp}{Yield under control condition}

\item{Ys}{Yield under stress condition}
}
\value{
TOL
}
\description{
Higher the TOL value indicates the genotype is tolerant to stress
}
\examples{
tol_index(500, 350)
}
\references{
Hossain, A. B. S., Sears, R. G., Cox, T. S., & Paulsen, G. M. (1990). Desiccation tolerance and its relationship to assimilate partitioning in winter wheat. Crop Science, 30(3), 622-627.
}
