% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{Y_index}
\alias{Y_index}
\title{Yield index}
\usage{
Y_index(Ys, Ms)
}
\arguments{
\item{Ys}{Yield under stress condition}

\item{Ms}{Mean Yield of all the genotypes under stress Condition}
}
\value{
Yield Index
}
\description{
Yield index
}
\examples{
Y_index(500, 300)
}
\references{
Gavuzzi, P., Rizza, F., Palumbo, M., Campanile, R. G., Ricciardi, G. L., & Borghi, B. (1997). Evaluation of field and laboratory predictors of drought and heat tolerance in winter cereals. Canadian journal of plant science, 77(4), 523-531.
}
