% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\docType{data}
\name{yield_data}
\alias{yield_data}
\title{Example data}
\format{
A data frame with 50 rows and 3 variables:
\describe{
\item{\code{Genotype}}{character Genotype}
\item{\code{Yp}}{integer Yield under control condition}
\item{\code{Ys}}{integer Yield under drought condition}
}
}
\source{
Simulated data, no external source were used
}
\usage{
yield_data
}
\description{
Yield data of rice in kg/ha under two different growth conditions
}
\references{
No external reference
}
\keyword{datasets}
