% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{YR_ratio}
\alias{YR_ratio}
\title{Yield reduction ratio (YR)}
\usage{
YR_ratio(Yp, Ys)
}
\arguments{
\item{Yp}{Yield under control condition}

\item{Ys}{Yield under stress condition}
}
\value{
YR
}
\description{
lesser the YR value more stable under stress conditions
}
\examples{
YR_ratio(500, 350)
}
