% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{YSI}
\alias{YSI}
\title{Yield reduction index or Yield Stability Index (YSI)}
\usage{
YSI(Yp, Ys)
}
\arguments{
\item{Yp}{Yield under control condition}

\item{Ys}{Yield under stress condition}
}
\value{
YSI
}
\description{
Higer YSI value depicts that particular genotype is stable under both normal and stressed conditions
}
\examples{
YSI(500, 350)
}
\references{
Bouslama, M., & Schapaugh Jr, W. T. (1984). Stress tolerance in soybeans. I. Evaluation of three screening techniques for heat and drought tolerance 1. Crop science, 24(5), 933-937.
}
