% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR-package.R
\docType{package}
\name{plantphysioR-package}
\alias{plantphysioR}
\alias{plantphysioR-package}
\title{plantphysioR: Fundamental Formulas for Plant Physiology}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions tailored for scientific and student communities involved in plant science research. Functionalities encompass estimation chlorophyll content, determination water potential of PEG6000 and functions related to estimation of yield related indices with respect to different growth conditions. Functions included in the packages are based on the formula from the available literature.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rameshram96/plantphysioR}
  \item Report bugs at \url{https://github.com/rameshram96/plantphysioR/issues}
}

}
\author{
\strong{Maintainer}: Ramesh Ramasamy \email{ramesh.rahu96@gmail.com} (\href{https://orcid.org/0000-0002-5433-1417}{ORCID})

}
\keyword{internal}
