% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR-package.R
\docType{package}
\name{plantphysioR-package}
\alias{plantphysioR}
\alias{plantphysioR-package}
\title{plantphysioR: Fundamental Formulas for Plant Physiology}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides some basic functions that may be useful for the scientific communities involved in the plant physiology research. This package includes functions for calculating chlorophyll and carotenoid content, calculating PEG600 requirement, and indices related to stress tolerance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rameshram96/plantphysioR}
  \item Report bugs at \url{https://github.com/rameshram96/plantphysioR/issues}
}

}
\author{
\strong{Maintainer}: Ramesh Ramasamy \email{ramesh.rahu96@gmail.com} (\href{https://orcid.org/0000-0002-5433-1417}{ORCID})

}
\keyword{internal}
