% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantphysioR.R
\name{yield_reduction}
\alias{yield_reduction}
\title{Yield Reduction
Claculate percent yield reduction over control}
\usage{
yield_reduction(Yp, Ys)
}
\arguments{
\item{Yp}{Yield under control condition}

\item{Ys}{Yield under stress condition}
}
\value{
YR
}
\description{
Yield Reduction
Claculate percent yield reduction over control
}
\examples{
yield_reduction(500, 350)
}
